Refactor ProductImageCard to be reused in product page. Update carousel to fit new card.
